% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{CallCentre_KPI}
\alias{CallCentre_KPI}
\title{Data CallCentre_KPI: Call Centre Key Performance Indicators}
\format{
A data frame with \emph{n} = 5007 observations and 11 variables:
\describe{
\item{ID_Call}{(int) call identification}
\item{Day}{(factor) weekday of the call (\code{Mon}, \code{Tue}, \code{Wed},..., \code{Sat})}
\item{Hour}{(num) hour of the call}
\item{Cust.Type}{(chr) customer's type (\code{Unknown}, \code{Private}, \code{Corporate})}
\item{Time.Auto}{(num) time (in seconds) spent interacting with the
automatic responder}
\item{Time.Queue}{(num) time (in seconds) spent waiting in the queue
for an operator}
\item{Time.Talk}{(num) duration (in seconds) of the (eventual)
conversation with an operator}
\item{Reason}{(chr) reason for contact (\code{Admin} -- administrative
issues; \code{Landline} -- issues with fixed line; \code{Mobile}; and
\code{Activ/Transf} -- calls related to the activation or the transfer of
a line)}
\item{Outcome}{(chr) outcome of the call (\code{Left.Aut_Resp} -- the client
hung up before entering the queue; \code{Left.Queue} -- the client hung up
while waiting for an operator; \code{Operator} -- the call was handled by
an operator)}
\item{Solved}{(chr) variable indicating whether the client's issue was
solved (\code{No}, \code{Yes}) -- available only for clients who interacted with
an operator}
\item{Satisf}{(factor) level of satisfaction expressed by customers
who interacted with an operator (\code{VLow}, \code{Low}, \code{Med}, \code{High}, \code{VHigh})}
}``
}
\usage{
data(CallCentre_KPI)
}
\description{
This data frame contains data on the calls to the call centre of a
telecommunication company. It contains information about the calls (day
and hour of the call), about the calling customer's characteristics
(private or corporate, or unknown), and about some key performance
indicators (KPI). In particular, it contains information on the duration
of the phases before the customer eventually reaches an operator. First,
an automated system directs incoming calls to the appropriate department
based on the reason for the call; then the customer waits in a queue until
an operator becomes available. It also contains a variable indicating
whether the caller hung up during a specific phase. For those customers
whose call was finally handled by an operator, we also know whether their
issue was solved and their level of satisfaction, expressed at the end of
the call.
}
\references{
Piccarreta, R., Tonini, D., & Trentini, F. (2026). \emph{From Data to
Decisions. An Introduction to Applied Statistics}. BUP. ISBN
9788823824096.
}
\keyword{datasets}
