% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{MktDATA}
\alias{MktDATA}
\title{Data MktDATA:}
\format{
{A data frame with \emph{n} = 2224 observations and 26 variables.}
}
\usage{
data(MktDATA)
}
\description{
This dataset is a modification of the original \code{\link{MktDATA.Orig}}
dataset and it is provided for user convenience.
}
\keyword{datasets}
