% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{Spending}
\alias{Spending}
\title{Data Spending: Household Income and Spending Patterns}
\format{
A data frame with \emph{n} = 751 observations and 4 variables:
\describe{
\item{Id}{(int) household identification number}
\item{Essentials}{(num) household monthly expenditures in essential
goods (e.g., groceries, utilities, basic transport); score
normalised to ensure comparability across household of different
sizes}
\item{Discretionary}{(num) household monthly expenditures in
discretionary goods (e.g., dining out, travel, electronics); score
normalised to ensure comparability across household of different
sizes}
\item{Income}{(num) monthly disposable income (in thousands of
pounds) per equivalised household (meaning adjusted to allow
meaningful comparisons across households with different size and
composition)}
}
}
\usage{
data(Spending)
}
\description{
The data frame contains information on a sample of households within a
specific demographic segment; it is used to explore the relationship
between monthly spending on \emph{Essentials} and \emph{Discretionary}
goods and monthly disposable \emph{Income}. All variables in the
data frame are normalised to ensure comparability across households of
different sizes. Income is expressed in thousands of pounds per month per
equivalised household to avoid scale artefacts.
}
\references{
Piccarreta, R., Tonini, D., & Trentini, F. (2026). \emph{From Data to
Decisions. An Introduction to Applied Statistics}. BUP. ISBN
9788823824096.
}
\keyword{datasets}
