% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{Transition}
\alias{Transition}
\title{Data Transition: Employee Productivity After Software Transition}
\format{
A data frame with \emph{n} = 130 observations and 4 variables:
\describe{
\item{Employee}{(int) employee id}
\item{Pre}{(num) employee's productivity with the old software}
\item{Post}{(num) employee's productivity with the new software
(after the training program)}
\item{Department}{(chr) employee's department (\code{Dept1}, \code{Dept2})}
}
}
\usage{
data(Transition)
}
\description{
The data frame refers to data on the employees of a company that adopted a
new management software system to replace its legacy tools. To facilitate
a smooth transition, all administrative staff will participate in a
structured training program designed to build proficiency with the new
system. As part of the rollout, employees from two departments have been
selected to complete the training first. The dataset contains information
on each (trained) employee's productivity using the old and the new
software, recorded after a defined period of use.
}
\references{
Piccarreta, R., Tonini, D., & Trentini, F. (2026). \emph{From Data to
Decisions. An Introduction to Applied Statistics}. BUP. ISBN
9788823824096.
}
\keyword{datasets}
