% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{Banner}
\alias{Banner}
\title{Data Banner: Banner Click-Through Rate Experiment}
\format{
A data frame with \emph{n} = 5976 observations and 3 variables:
\describe{
\item{User}{(int) user id}
\item{Banner}{(chr) banner shown to the user (\code{Original}, \code{Redesigned})}
\item{Click}{(logical) logical vector indicating whether the user
clicked on the banner (TRUE) or not}
}
}
\usage{
data(Banner)
}
\description{
The data frame contains data collected by a company interested in comparing
a \emph{Redesigned} banner (featuring a brighter colour scheme and modified
layout) with the \emph{Original} version. A random sample of users from a
social network is shown one of two banners, and the effectiveness of the
two banners is assessed via the \strong{click-through rate (CTR)} -- the
proportion of users who click on a banner after viewing it.
}
\references{
Piccarreta, R., Tonini, D., & Trentini, F. (2026). \emph{From Data to
Decisions. An Introduction to Applied Statistics}. BUP. ISBN
9788823824096.
}
\keyword{datasets}
