% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{Banner}
\alias{Banner}
\title{Data: Banner}
\format{
{
A dataframe with \emph{n} = 2350 observations and the following 3 variables
(levels of the variables listed in alphabetical order):
\itemize{
\item \strong{User} (\code{int}): user id
\item \strong{Banner} (\code{chr}): banner shown to the user (Original, Redesigned)
\item \strong{Click} (\code{logical}): logical vector indicating whether the user
clicked on the banner (\code{TRUE}) or not
}
}
}
\usage{
data(Banner)
}
\description{
The dataframe contains data collected by a company interested in comparing a
Redesigned banner (featuring a brighter colour scheme and modified layout)
with the Original version. A random sample of users from a social network is
shown one of two banners, and the effectiveness of the two banners is
assessed via the click-through rate (CTR) -- the proportion of users who click
on a banner after viewing it.
}
\keyword{datasets}
