% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{BasketValue}
\alias{BasketValue}
\title{Data BasketValue: Supermarket Spending and Basket Variety}
\format{
A data frame with \emph{n} = 128 observations and 5 variables:
\describe{
\item{Id}{(int) customer identification number}
\item{Age}{(num) age of the customers in years}
\item{Wealth}{(num) composite indicator of household economic
well-being obtained by estimating household wealth from the average
market price and rental value of dwellings located in the same street
as each customer's declared address, and adjusting for the number of
cohabitants}
\item{Promo}{(factor) binary variable indicating whether promotional
products are frequently purchased (No, Yes)}
\item{SpendMix}{(num) composite index which reflects both spending
level and basket variety}
}
}
\usage{
data(BasketValue)
}
\description{
The data frame contains data on a supermarket customers living relatively
close to the store. Based on the details provided when the loyalty card
was registered and on recorded expenses, for each customer information is
available on a composite index, \emph{SpendMix}, which captures how
supermarket clients differ not only in how much they spend but also in the
range of products they purchase. We are interested to identify factors
associated with the level of spending and basket variety. Some
characteristics (e.g., income and age) are obvious candidates: richer
individuals usually spend more, and spending habits often change over the
course of a person's life. Yet behavioural differences can also play a
role. For instance, shoppers who are particularly responsive to discounts
may be more inclined to experiment with new items or brands.
}
\references{
Piccarreta, R., Tonini, D., & Trentini, F. (2026). \emph{From Data to
Decisions: An Applied Introduction to Statistics}. BUP. ISBN
9788823824096.
}
\keyword{datasets}
