% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBStats_Main_Inference.R
\name{CI.diffmean}
\alias{CI.diffmean}
\title{Confidence intervals for the mean difference}
\usage{
CI.diffmean(
  x,
  y,
  type = "independent",
  sigma.x = NULL,
  sigma.y = NULL,
  sigma.by = NULL,
  sigma.d = NULL,
  conf.level = 0.95,
  digits = 2,
  by,
  var.test = FALSE,
  data,
  ...
)
}
\arguments{
\item{x, y}{Quoted or unquoted strings identifying the variables whose
distribution has to be analysed. \code{x} and \code{y} can be the
name of a vector or a factor in the workspace or the name of one of
the columns in the data frame specified in the \code{data} argument.
It is possible to use a mixed specification (e.g, one vector and one
column in data).}

\item{type}{A length-one character vector specifying the type of samples.
Allowed values are \code{"independent"} or \code{"paired"}.}

\item{sigma.x, sigma.y}{Optional numeric values representing the population
standard deviations for the two population. If \code{NULL} they estimate
the populations standard deviations using the data.}

\item{sigma.by}{An optional numeric value used to specify the possibly known
variances for the two independent samples identified via \code{by}.
\code{sigma.by} can be a single value indicating the same variance in the
two by-groups, or a vector with two values, specifying the variances in
the two by-groups. To avoid errors, in the latter case the vector should be
named, with names coinciding with the two levels of \code{by}.}

\item{sigma.d}{An optional numeric value that specify the possibly known
variance of the difference in \bold{paired} samples.}

\item{conf.level}{Numeric value corresponding to the confidence level
required. Default to 0.95.}

\item{digits}{Integer value specifying the number of decimals used
to round statistics; default to 2.}

\item{by}{Quoted or unquoted string identifying the variable used to
identify the two samples whose means have to be compared. Note that
only \bold{one} between \code{y} and \code{by} can be specified. More
specifically:
\itemize{
\item \code{x}, \code{y} should be specified when data on the two sub-samples
are stored in two distinct vectors,
\item \code{x}, \code{by} should be specified when data on the two sub-samples
should be obtained by splitting \code{x} into two groups based on the two
values of the by vector; in this case, \code{x} and \code{by} must have
the same length, \code{x} must be numeric, and \code{by} can take only
two values. Note that this option is available only for \bold{independent}
samples.
}}

\item{var.test}{Logical value indicating whether to run a test on the
equality of variance in two (\bold{independent}) samples or not (default).}

\item{data}{An optional data frame containing the variable to be analysed.
If not found in \code{data}, the variables are taken from the environment
from which \code{CI.diffmean()} is called.}

\item{...}{Additional arguments to be passed to low level functions.}
}
\value{
A table reporting the confidence intervals for the population mean
difference.
}
\description{
\code{CI.diffmean()} computes confidence intervals for the difference
between the means of two populations.
}
\examples{
data(MktDATA, package = "UBStats")

# CI for the difference between means in groups of males and females
# based on x, y: build vectors with data on the two groups
AOV_F <- MktDATA$AOV[MktDATA$Gender == "F"]
AOV_M <- MktDATA$AOV[MktDATA$Gender == "M"]
CI.diffmean(x = AOV_M, y = AOV_F, type = "independent")

# Same as above but using x, by: x split based on the by's (2) levels
MktDATA$Gender.R <- factor(MktDATA$Gender, levels = c("M", "F"))
CI.diffmean(x = AOV, by = Gender.R, type = "independent", data = MktDATA)

}
\seealso{
\code{\link{TEST.diffmean}()} for computing a test on the difference
between two population means.
}
\author{
Raffaella Piccarreta \email{raffaella.piccarreta@unibocconi.it}
}
