% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBStats_Main_Inference.R
\name{CI.diffprop}
\alias{CI.diffprop}
\title{Confidence intervals for the proportion difference}
\usage{
CI.diffprop(
  x,
  y,
  success.x = NULL,
  success.y = NULL,
  conf.level = 0.95,
  digits = 2,
  by,
  data,
  ...
)
}
\arguments{
\item{x, y}{Unquoted strings identifying the variables whose
distribution has to be analysed. \code{x} and \code{y} can be the
name of a vector or a factor in the workspace or the name of one of
the columns in the data frame specified in the \code{data} argument.
It is possible to use a mixed specification (e.g, one vector and one
column in data).}

\item{success.x, success.y}{If \code{x,y} are factors, character
vectors, or numeric non-binary vectors, success must be used to indicate
the category/value corresponding to success in the populations. These
arguments can be omitted (\code{NULL}, default) if \code{x,y} are binary
numeric vectors (taking values 0 or 1 only; in this case success is
assumed to correspond to 1) or a logical vector (in these cases success
is assumed to correspond to \code{TRUE}).}

\item{conf.level}{Numeric value corresponding to the confidence level
required. Default to 0.95.}

\item{digits}{Integer value specifying the number of decimals used
to round statistics; default to 2.}

\item{by}{Unquoted string identifying the variable used to
identify the two samples whose means have to be compared. Note that
only \bold{one} between \code{y} and \code{by} can be specified. More
specifically:
\itemize{
\item \code{x}, \code{y} should be specified when data on the two sub-samples
are stored in two distinct vectors,
\item \code{x}, \code{by} should be specified when data on the two sub-samples
should be obtained by splitting \code{x} into two groups based on the two
values of the by vector; in this case, \code{x} and \code{by} must have
the same length, \code{x} must be numeric, and \code{by} can take only
two values. Note that this option is available only for \bold{independent}
samples.
}}

\item{data}{An optional data frame containing the variable to be analysed.
If not found in \code{data}, the variables are taken from the environment
from which \code{CI.diffprop()} is called.}

\item{...}{Additional arguments to be passed to low level functions.}
}
\value{
A table reporting the confidence intervals for the population
proportion difference.
}
\description{
\code{CI.diffprop()} computes confidence intervals for the difference
between the proportion of successes of two populations.
}
\examples{
data(MktDATA, package = "UBStats")

# Based on x,y: build vectors with data on the two groups
WouldSuggest_F <- MktDATA$WouldSuggest[MktDATA$Gender == "F"]
WouldSuggest_M <- MktDATA$WouldSuggest[MktDATA$Gender == "M"]
CI.diffprop(x = WouldSuggest_M, y = WouldSuggest_F, success.x = "Yes",
  conf.level = 0.99, digits = 2)

# CI for difference between proportion based on logical variable
NoChildren <- (MktDATA$Children == 0)
Deals_w_child <- MktDATA$Deals.ge50[NoChildren == FALSE]
Deals_no_child <- MktDATA$Deals.ge50[NoChildren == TRUE]
CI.diffprop(x = Deals_w_child, y = Deals_no_child, conf.level = 0.9,
  digits = 3)

}
\seealso{
\code{\link{TEST.diffprop}()} for computing a test on the difference
between two population proportions of success.
}
\author{
Raffaella Piccarreta \email{raffaella.piccarreta@unibocconi.it}
}
