% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBStats_Main_Inference.R
\name{CI.mean}
\alias{CI.mean}
\title{Confidence intervals for the mean}
\usage{
CI.mean(x, sigma = NULL, conf.level = 0.95, digits = 2, data, ...)
}
\arguments{
\item{x}{An unquoted string identifying the variable whose
distribution has to be analysed. \code{x} can be the name of a vector
or a factor in the workspace or the name of one of the columns in the
data frame specified in the \code{data} argument.}

\item{sigma}{An optional numeric value representing the population standard
deviation. If \code{NULL} it estimates the population standard using the
data.}

\item{conf.level}{Numeric value corresponding to the confidence level
required. Default to 0.95.}

\item{digits}{Integer value specifying the number of decimals used
to round statistics; default to 2.}

\item{data}{An optional data frame containing the variable to be analysed.
If not found in \code{data}, the variables are taken from the environment
from which \code{CI.mean()} is called.}

\item{...}{Additional arguments to be passed to low level functions.}
}
\value{
A table reporting the confidence intervals for the population mean.
}
\description{
\code{CI.mean()} computes confidence intervals for the mean of a population.
}
\examples{
data(MktDATA, package = "UBStats")

# CI for the mean with KNOWN variance with default options
CI.mean(AOV, sigma = 30, data = MktDATA)

# CI for the mean with UNKNOWN variance,  confidence level 0.99
CI.mean(AOV, conf.level = 0.99, data = MktDATA)

}
\seealso{
\code{\link{TEST.mean}()} for computing a test on the population
mean.
}
\author{
Raffaella Piccarreta \email{raffaella.piccarreta@unibocconi.it}
}
