% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBStats_Main_Inference.R
\name{CI.prop}
\alias{CI.prop}
\title{Confidence intervals for the proportion}
\usage{
CI.prop(x, success = NULL, conf.level = 0.95, digits = 2, data, ...)
}
\arguments{
\item{x}{A quoted or unquoted string identifying the variable whose
distribution has to be analysed. \code{x} can be the name of a vector
or a factor in the workspace or the name of one of the columns in the
data frame specified in the \code{data} argument.}

\item{success}{If \code{x} is a factor, a character vector, or a numeric
non-binary vector, success must be used to indicate the category/value
corresponding to success. The argument can be omitted (\code{NULL},
default) if \code{x} is a binary numeric vector (takes values 0 or 1 only;
in this case success is assumed to be 1) or a logical vector (in these
cases success is assumed to be \code{TRUE}).}

\item{conf.level}{Numeric value corresponding to the confidence level
required. Default to 0.95.}

\item{digits}{Integer value specifying the number of decimals used
to round statistics; default to 2.}

\item{data}{An optional data frame containing the variable to be analysed.
If not found in \code{data}, the variables are taken from the environment
from which \code{CI.prop()} is called.}

\item{...}{Additional arguments to be passed to low level functions.}
}
\value{
A table reporting the confidence intervals for the population
proportion of successes.
}
\description{
\code{CI.prop()} computes confidence intervals for the proportion of
successes for a population.
}
\examples{
data(MktDATA, package = "UBStats")

# Character: a specific value
CI.prop(WouldSuggest, success = "Yes", data = MktDATA)

# Factor: a specific level
CI.prop(Education, success = "Post-Grad", conf.level = 0.9,
  digits = 4, data = MktDATA)

# Numeric vector: a specific value
CI.prop(Children, success = 2, conf.level = 0.99, data = MktDATA)

# CI for the proportion based on a binary variable (taking values 0/1)
CI.prop(LastCampaign, conf.level = 0.9, digits = 3, data = MktDATA)

# Build a (logical) vector indicating whether a condition is satisfied
IsTop <- MktDATA$CustClass == "Gold" | MktDATA$CustClass == "Platinum"
CI.prop(IsTop, conf.level = 0.9, digits = 3, data = MktDATA)

}
\seealso{
\code{\link{TEST.prop}()} for computing a test on the proportion
of successes.
}
\author{
Raffaella Piccarreta \email{raffaella.piccarreta@unibocconi.it}
}
