% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBStats_Main_TablesPlots.R
\name{distr.plot.xy}
\alias{distr.plot.xy}
\title{Analysis of a bivariate distribution using plots}
\usage{
distr.plot.xy(
  x,
  y,
  freq = "Counts",
  freq.type = "joint",
  plot.type,
  bar.type = "stacked",
  fitline = FALSE,
  legend = TRUE,
  breaks.x,
  breaks.y,
  adj.breaks = TRUE,
  interval.x = FALSE,
  interval.y = FALSE,
  bw = FALSE,
  color = NULL,
  data,
  ...
)
}
\arguments{
\item{x, y}{Unquoted strings identifying the variables whose
distribution has to be analysed. \code{x} and \code{y} can be the
name of a vector or a factor in the workspace or the name of one of
the columns in the data frame specified in the \code{data} argument.
Note that in the output \code{x} represents the \emph{row} variable
while \code{y} corresponds to the \emph{column} variable.}

\item{freq}{A string providing the frequencies to be displayed. Allowed
values (possibly abbreviated) are \code{"counts"},
\code{"percentages"} and \code{"proportions"}.}

\item{freq.type}{A length-one character vector providing the type of
frequencies to be displayed. Allowed options are \code{joint} (default)
for joint frequencies, \code{x|y} (or \code{column}) for the distributions
of \code{x} conditioned to \code{y}, and \code{y|x} (or \code{row}) for
the distributions of \code{y} conditioned to \code{x}.}

\item{plot.type}{A length-one character vector providing the plot to generate.
Allowed values are \code{"bars"}, \code{"scatter"}, and
\code{"boxplot"}.}

\item{bar.type}{A length-one character vector indicating whether in a bar plot,
stacked (\code{bar.type = "stacked"}, default) or side-by-side
(\code{bar.type = "beside"}) bars should be displayed.}

\item{fitline}{Logical value indicating whether the line of best fit (also
called trendline or regression line) should be added to a scatterplot
(\code{fitline = TRUE}) or not (\code{fitline = FALSE}; default).}

\item{legend}{Logical value indicating whether a legend should be displayed
in the plot (\code{legend = TRUE}; default) or not (\code{legend = FALSE}).}

\item{breaks.x}{Allows to classify a \emph{numerical} variable \code{x} into
intervals. It can be an integer specifying the number of intervals of
equal width used to classify \code{x}, or a vector of increasing numeric
values defining the endpoints of intervals (closed on the left and open
on the right; the last interval is closed on the right too). To cover
he entire range of values the maximum and the minimum values should be
included between the first and the last break. It is possible to specify
a set of breaks covering only a portion of the \code{x} range.}

\item{breaks.y}{Allows to classify a \emph{numerical} variable \code{y} into
intervals. It can be an integer specifying the number of intervals of
equal width used to classify \code{y}, or a vector of increasing numeric
values defining the endpoints of intervals (closed on the left and open
on the right; the last interval is closed on the right too). To cover
he entire range of values the maximum and the minimum values should be
included between the first and the last break. It is possible to specify
a set of breaks covering only a portion of the \code{y} range.}

\item{adj.breaks}{Logical value indicating whether numbers displayed
should avoid using scientific notation; default to \code{TRUE}.}

\item{interval.x}{Logical value indicating whether \code{x} is a variable
measured in classes (default \code{TRUE}). If the detected intervals are not
consistent (e.g. overlapping intervals, or intervals with upper endpoint
higher than the lower one), the variable is tabulated as it is, even if
results are not necessarily consistent.}

\item{interval.y}{Logical value indicating whether \code{y} is a variable
measured in classes (default \code{TRUE}). If the detected intervals are not
consistent (e.g. overlapping intervals, or intervals with upper endpoint
higher than the lower one), the variable is tabulated as it is, even if
results are not necessarily consistent.}

\item{bw}{Logical value indicating whether plots are coloured using a standard palette
(\code{FALSE}) rather than in scale of greys.}

\item{color}{String vector indicating the specific colors to use in the plot, or
\code{NULL} (defualt) which indicates to use a standard palette.}

\item{data}{An optional data frame containing the variable to be analysed.
If not found in \code{data}, the variables are taken from the environment
from which \code{distr.plot.xy()} is called.}

\item{...}{Additional arguments to be passed to low level functions.}
}
\description{
\code{distr.plot.xy()} generates plots of a bivariate distribution.
}
\examples{
data(MktDATA, package = "UBStats")

# Bivariate bar plots
distr.plot.xy(CustClass, Children, freq = "Counts", freq.type = "joint",
  plot.type = "bars", data = MktDATA)
distr.plot.xy(TotPurch, Income, freq = "Percentages", freq.type = "x|y",
  plot.type = "bars", breaks.x = c(0, 5, 10, 15, 20, 35),
  interval.y = TRUE, data = MktDATA)

# Side-by-side boxplots
distr.plot.xy(x = AOV, y = Education, plot.type = "boxplot",
  freq.type = "y|x", data = MktDATA)

# Scatterplots
distr.plot.xy(Baseline, TotVal, plot.type = "scatter", fitline = TRUE,
  data = MktDATA)

# Bubble plots
#  - two characters/factors
distr.plot.xy(Education, LikeMost, plot.type = "scatter", data = MktDATA)
#  - classified variables (i.e. not properly numerical)
distr.plot.xy(Income.S, TotPurch, interval.x = TRUE,
  breaks.y = c(0, 5, 10, 15, 20, 35), plot.type = "scatter", data = MktDATA)

}
\seealso{
\code{\link{distr.table.xy}()} for computing a bivariate
distribution.

\code{\link{distr.table.x}()} for computing a univariate
distribution.

\code{\link{distr.plot.x}()} for plotting a univariate
distribution.
}
\author{
Raffaella Piccarreta \email{raffaella.piccarreta@unibocconi.it}
}
