% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBStats_Main_TablesPlots.R
\name{distr.summary.x}
\alias{distr.summary.x}
\title{Summary statistics for a single variable}
\usage{
distr.summary.x(
  x,
  by1,
  by2,
  stats = c("summary"),
  digits = 2,
  f.digits = 4,
  data,
  ...
)
}
\arguments{
\item{x}{An unquoted string identifying the variable whose
distribution has to be analysed. \code{x} can be the name of a vector
or a factor in the workspace or the name of one of the columns in the
data frame specified in the \code{data} argument.}

\item{by1, by2}{Unquoted strings identifying optional variables
(typically taking few values/levels) used to build conditional summaries,
that can be defined same way as x.}

\item{stats}{A character vector providing the summary statistics to compute
(more summaries can be specified).
The following options are available:
\itemize{
\item \code{"summary"} returns min, q1 , median, mean, q3, max, sd, var,
\item \code{"central"} returns central tendency measures only,
\item \code{"dispersion"} returns measures of dispersion only,
\item \code{"fivenumbers"} returns the five number summary,
\item \code{"quartiles"}, \code{"quintiles"}, \code{"deciles"},
\code{"percentiles"} return the corresponding quantities.
}

It is also possible to list the following specific statistics:
\code{"q1"}, \code{"q2"}, \code{"q3"}, \code{"mean"}, \code{"median"},
\code{"mode"} (which returns the mode, the number of modes and the
proportion of cases with modal value respectively), \code{"min"},
\code{"max"}, \code{"sd"}, \code{"var"}, \code{"cv"} (coefficient of
variation), \code{"range"}, \code{"IQrange"}, or \code{"p1"},
\code{"p2"},..., \code{"p100"} (i.e. specific percentiles).}

\item{digits}{Integer value specifying the number of decimals used
to round summary statistics; default to 4.}

\item{f.digits}{Integer value specifying the number of decimals used
to round frequencies; default to 2.}

\item{data}{An optional data frame containing the variable to be analysed.
If not found in \code{data}, the variables are taken from the environment
from which \code{distr.summary.x()} is called.}

\item{...}{Additional arguments to be passed to low level functions.}
}
\value{
A table listing the values taken by the variable arranged in standard
order (logical, alphabetical or numerical order for vectors, order of levels
for factors), and the frequencies requested.
}
\description{
\code{distr.summary.x()} computes the frequency table of a vector or a factor.
}
\examples{
data(MktDATA, package = "UBStats")

# Default summaries
distr.summary.x(x = AOV, data = MktDATA)
distr.summary.x(x = AOV, stats = c("central", "dispersion", "fivenumbers"),
  data = MktDATA)
distr.summary.x(x = AOV, stats = c("mode", "mean", "sd", "cv", "fivenumbers"),
  data = MktDATA)

# Conditional summaries
distr.summary.x(x = TotVal, by1 = Gender,
  stats = c("p5", "p10", "p25", "p50", "p75", "p90", "p95"),
  digits = 1, data = MktDATA)
distr.summary.x(x = AOV, by1 = Gender, by2 = Kids,
  stats = "fivenumbers", data = MktDATA)

}
\seealso{
\code{\link{distr.table.x}()} for computing a univariate
distribution.

\code{\link{distr.plot.x}()} for plotting a univariate
distribution.
}
\author{
Raffaella Piccarreta \email{raffaella.piccarreta@unibocconi.it}
}
