% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBStats_Main_TablesPlots.R
\name{distr.table.x}
\alias{distr.table.x}
\title{Analysis of a univariate distribution using tables}
\usage{
distr.table.x(
  x,
  freq = c("counts", "proportions"),
  total = TRUE,
  breaks,
  adj.breaks = TRUE,
  interval = FALSE,
  f.digits = 2,
  p.digits = 0,
  d.digits = 5,
  data,
  ...
)
}
\arguments{
\item{x}{A quoted or unquoted string identifying the variable whose
distribution has to be analysed. \code{x} can be the name of a vector
or a factor in the workspace or the name of one of the columns in the
data frame specified in the \code{data} argument.}

\item{freq}{A string providing the type of frequencies to be displayed.
Allowed values (possibly abbreviated) are \code{"counts"},
\code{"percentages"}, \code{"proportions"}, \code{"densities"}
(only for variables classified into intervals), and \code{"cumulative"}.
If no frequency is specified, \code{"counts"} and \code{"proportions"}
are displayed by default if only \code{"cumulative"} is requested,
counts and proportions will also be displayed, with their respective
cumulative frequencies.}

\item{total}{Logical value indicating whether the sum of the requested
frequencies should be added to the table; default to \code{TRUE}.}

\item{breaks}{Allows to classify a \emph{numerical} variable \code{x} into
intervals. It can be an integer specifying the number of intervals of
equal width used to classify \code{x}, or a vector of increasing numeric
values defining the endpoints of intervals (closed on the left and open
on the right; the last interval is closed on the right too). To cover
he entire range of values the maximum and the minimum values should be
included between the first and the last break. It is possible to specify
a set of breaks covering only a portion of the \code{x} range.}

\item{adj.breaks}{Logical value indicating whether numbers displayed
should avoid using scientific notation; default to \code{TRUE}.}

\item{interval}{Logical value indicating whether \code{x} is a variable
measured in intervals (\code{TRUE}). If the detected intervals are not
consistent (e.g. overlapping intervals, or intervals with upper endpoint
higher than the lower one), the variable is tabulated as it is, even if
results are not necessarily consistent.}

\item{f.digits}{Integer value specifying the number of decimals used
to round proportions; default to 2.}

\item{p.digits}{Integer value specifying the number of decimals used
to round percentages; default to 0.}

\item{d.digits}{Integer value specifying the number of decimals used
to round densities; default to 5.}

\item{data}{An optional data frame containing the variable to be analysed.
If not found in \code{data}, the variables are taken from the environment
from which \code{distr.table.x()} is called.}

\item{...}{Additional arguments to be passed to low level functions.}
}
\value{
A table listing the values taken by the variable arranged in standard
order (logical, alphabetical or numerical order for vectors, order of levels
for factors), and the frequencies requested.
}
\description{
\code{distr.table.x()} computes the frequency table of a vector or a factor.
}
\examples{
data(MktDATA, package = "UBStats")

distr.table.x(Education, data = MktDATA)

distr.table.x(Children, data = MktDATA, freq = c("count", "prop", "cum"),
  f.digits = 4)

distr.table.x(AOV, breaks = 6, freq = c("Count", "Perc", "Cum"),
  p.digits = 2, data = MktDATA)

distr.table.x(AOV, breaks = c(0, 20, 30, 50, 100, 180),
  freq = c("Count", "Perc", "Cum", "Densities"), p.digits = 2,
  data = MktDATA)

}
\seealso{
\code{\link{distr.plot.x}()} for plotting a univariate
distribution.

\code{\link{distr.table.xy}()} for computing a bivariate
distribution.

\code{\link{distr.plot.xy}()} for plotting a bivariate
distribution.
}
\author{
Raffaella Piccarreta \email{raffaella.piccarreta@unibocconi.it}
}
