% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBStats_Main_TablesPlots.R
\name{distr.table.xy}
\alias{distr.table.xy}
\title{Analysis of a bivariate distribution using tables}
\usage{
distr.table.xy(
  x,
  y,
  freq = "counts",
  freq.type = "joint",
  total = TRUE,
  breaks.x,
  breaks.y,
  adj.breaks = TRUE,
  interval.x = FALSE,
  interval.y = FALSE,
  f.digits = 2,
  p.digits = 0,
  data,
  ...
)
}
\arguments{
\item{x, y}{Unquoted strings identifying the variables whose
distribution has to be analysed. \code{x} and \code{y} can be the
name of a vector or a factor in the workspace or the name of one of
the columns in the data frame specified in the \code{data} argument.
Note that in the table \code{x} is displayed on the \emph{rows} and
\code{y} on the \emph{columns}.}

\item{freq}{A single character providing the frequencies to be displayed.
Allowed values (possibly abbreviated) are \code{"counts"},
\code{"percentages"} and \code{"proportions"}.}

\item{freq.type}{A single character providing the type of
frequencies to be displayed. Allowed options are \code{joint} (default)
for joint frequencies, \code{x|y} (or \code{column}) for the distributions
of \code{x} conditioned to \code{y}, and \code{y|x} (or \code{row}) for
the distributions of \code{y} conditioned to \code{x}.}

\item{total}{Logical value indicating whether the sum of the requested
frequencies should be added to the table; default to \code{TRUE}.}

\item{breaks.x}{Allows to classify a \emph{numerical} variable \code{x} into
intervals. It can be an integer specifying the number of intervals of
equal width used to classify \code{x}, or a vector of increasing numeric
values defining the endpoints of intervals (closed on the left and open
on the right; the last interval is closed on the right too). To cover
he entire range of values the maximum and the minimum values should be
included between the first and the last break. It is possible to specify
a set of breaks covering only a portion of the \code{x} range.}

\item{breaks.y}{Allows to classify a \emph{numerical} variable \code{y} into
intervals. It can be an integer specifying the number of intervals of
equal width used to classify \code{y}, or a vector of increasing numeric
values defining the endpoints of intervals (closed on the left and open
on the right; the last interval is closed on the right too). To cover
he entire range of values the maximum and the minimum values should be
included between the first and the last break. It is possible to specify
a set of breaks covering only a portion of the \code{y} range.}

\item{adj.breaks}{Logical value indicating whether numbers displayed
should avoid using scientific notation; default to \code{TRUE}.}

\item{interval.x}{Logical value indicating whether \code{x} is a variable
measured in classes (default \code{TRUE}). If the detected intervals are not
consistent (e.g. overlapping intervals, or intervals with upper endpoint
higher than the lower one), the variable is tabulated as it is, even if
results are not necessarily consistent.}

\item{interval.y}{Logical value indicating whether \code{y} is a variable
measured in classes (default \code{TRUE}). If the detected intervals are not
consistent (e.g. overlapping intervals, or intervals with upper endpoint
higher than the lower one), the variable is tabulated as it is, even if
results are not necessarily consistent.}

\item{f.digits}{Integer value specifying the number of decimals used
to round proportions; default to 2.}

\item{p.digits}{Integer value specifying the number of decimals used
to round percentages; default to 0.}

\item{data}{An optional data frame containing the variable to be analysed.
If not found in \code{data}, the variables are taken from the environment
from which \code{distr.table.xy()} is called.}

\item{...}{Additional arguments to be passed to low level functions.}
}
\value{
A table listing the values taken by the variables arranged in standard
order (logical, alphabetical or numerical order for vectors, order of levels
for factors), and the frequencies requested.
}
\description{
\code{distr.table.xy()} allows displaying tables of joint or conditional
distributions.
}
\examples{
data(MktDATA, package = "UBStats")

# Character vectors, factors, and discrete numeric vectors
distr.table.xy(LikeMost, Children, data = MktDATA) # default: joint counts

# Conditional distribution (proportions) of x|y
distr.table.xy(LikeMost, Education, freq = "Prop", freq.type = "x|y",
  data = MktDATA)

# Conditional distribution (proportions) of x|y
distr.table.xy(LikeMost, Education, freq = "Prop", freq.type = "x|y",
  data = MktDATA)

# Conditional distribution (\%) of y|x (columns-conditionals)
distr.table.xy(CustClass, Children, freq = "Percentages", freq.type = "row",
  data = MktDATA)

# A numerical variable classified into intervals and a factor
distr.table.xy(CustClass, TotPurch, breaks.y = c(0, 5, 10, 15, 20, 35),
  freq = "Prop", freq.type = "y|x", data = MktDATA)

# A numerical variable classified into intervals and a factor
distr.table.xy(CustClass, TotPurch, breaks.y = c(0, 5, 10, 15, 20, 35),
  freq = "Prop", freq.type = "y|x", data = MktDATA,
  msg.control = list(err = FALSE, warn = FALSE, msg = TRUE))

# Two numerical  variables classified into intervals
distr.table.xy(Income.S, TotPurch, interval.x = TRUE,
  breaks.y = c(0, 5, 10, 15, 20, 35),
  freq = "Counts", freq.type = "row", data = MktDATA)

# Two numerical classified variables
distr.table.xy(Income.S, TotPurch, interval.x = TRUE,
  breaks.y = c(0, 5, 10, 15, 20, 35),
  freq = "Counts", freq.type = "row", data = MktDATA,
  msg.control = list(err = FALSE, warn = FALSE, msg = TRUE))

}
\seealso{
\code{\link{distr.plot.xy}()} for plotting a bivariate
distribution.

\code{\link{distr.table.x}()} for computing a univariate
distribution.

\code{\link{distr.plot.x}()} for plotting a univariate
distribution.
}
\author{
Raffaella Piccarreta \email{raffaella.piccarreta@unibocconi.it}
}
