% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{Grocery_NE}
\alias{Grocery_NE}
\title{Data Grocery_NE: Grocery Chain Customer Survey (North-East Italy)}
\format{
A data frame with \emph{n} = 3114 observations and 26 variables:
\describe{
\item{Id}{(int) customer identification}
\item{Sex}{(chr) customer's sex at birth (\code{F} or \code{M})}
\item{Age}{(num) customer's age (in years)}
\item{Tenure}{(factor) Customer tenure in years (coded in classes:
\verb{[0,1)}, \verb{[1,3)}, \verb{[3,6)}, \verb{[6,10)}, \verb{[10,15)}, \verb{[15,25)}, \verb{[25,35)})}
\item{Status}{(chr) customers' status (\code{Active}, \code{Silent})}
\item{FavShop}{(chr) store customers visit the most (\code{NE.01}, \code{NE.02},
..., \code{NE.07})}
\item{FavShop_Region}{(chr) macro region where the favourite shop is
located (here, only \code{North-East})}
\item{TotShops}{(num) number of stores visited by the customer}
\item{WeekDay}{(chr) preferred shopping day (\code{1:Mon}, \code{2:Tue}, ..., \code{7:Sun})}
\item{TimeSlot}{(chr) preferred shopping time slot (\code{08-12}, \code{12-14},
\code{14-17}, \code{17-23})}
\item{Satisf}{(chr) Customer's overall declared satisfaction with the
retailer (\code{VLow}, \code{Low}, \code{QLow} -- that is, moderately low; \code{Med},
\code{QHigh} -- that is, moderately high; \code{High}, \code{VHigh})}
\item{Complaint}{(factor) major weakness of the typically visited
store (\code{Quality&Variety}, \code{Prices}, \code{Resupply}, \code{Staff}, \code{Crowded},
\code{Checkout})}
\item{NMonths}{(num) number of months in which the customers visited
a store at least once}
\item{MonthExp}{(chr) amount spent per month, in classes (\verb{[0,50)},
\verb{[50,100)}, \verb{[100,150)}, \verb{[150,200)}, \verb{[200,300)}, \verb{[300,400)},
\verb{[400,600)}, \verb{[600,800]})}
\item{Transact_M}{(num) transactions per month}
\item{TBP}{(num) time between purchases}
\item{TotExp}{(num) amount spent in the last 12 months}
\item{TotVisits}{(num) total number of visits in the last year}
\item{Receipt}{(num) average receipt (transaction value)}
\item{Visits_Regular}{(num) score (ranging from 0 to 100) indicating
the regularity of customer's shopping trips}
\item{Spending_Regular}{(num) score (ranging from 0 to 100)
indicating the regularity of customer's spending}
\item{Discount}{(num) average discount on purchased products}
\item{CrossSelling}{(num) index measuring how diverse a customer's
purchases are across different product categories}
\item{RecencyScore}{(num) index reflecting how recently a customer
last interacted with a company (higher scores reflecting more
recent interactions)}
\item{MonetaryScore}{(num) index reflecting the overall value of the
customer to the business}
\item{FrequencyScore}{(num) index reflecting the regularity or
repetition of customer transactions, with higher scores indicating
more frequent purchases}
}
}
\usage{
data(Grocery_NE)
}
\description{
This dataset contains the variables from a survey on a sample of
customers from a grocery chain operating in Italy. Specifically, data
refer to the sub-sample of customers who regularly shop in stores located
in north-eastern Italian region. Information is available on the activity
observed in the last year (number of visits and transactions, amount
spent), on customers' satisfaction with the retailer, and on the
perceived weakness of the most frequently visited store.
}
\references{
Piccarreta, R., Tonini, D., & Trentini, F. (2026). \emph{From Data to
Decisions. An Introduction to Applied Statistics}. BUP. ISBN
9788823824096.
}
\keyword{datasets}
