% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{JobEngage}
\alias{JobEngage}
\title{Data JobEngage: Employee Job Engagement Survey}
\format{
A data frame with \emph{n} = 256 observations and 11 variables:
\describe{
\item{Id}{(int) respondent id}
\item{Engagement}{(num) score reflecting the respondent's level of
engagement}
\item{Leadership}{(num) respondent evaluation of managers, as well as
the quality of their relationship with them}
\item{Team}{(num) respondent evaluation of co-workers, as well as
the quality of their relationship with them}
\item{Environment}{(num) respondent evaluation of the quality of the
work environment}
\item{Growth}{(num) respondent perceived evaluation of the opportunity
of professional development and advancement; it captures the
degree to which employees see clear paths to promotion within the
company and feel they are provided
with opportunities to learn and develop skills}
\item{Compensation}{(num) respondent perceived fairness of
compensation}
\item{WorkPlace}{(num) respondent evaluation of the physical work
environment, including aspects such as lighting, noise level, the
amount of space they have access to, and ergonomics}
\item{RoleConflict}{(num) extent to which employees feel that the
expectations connected with their role are inconsistent or
incompatible with each other (for example, when they receive
conflicting instructions from different supervisors or face
demands that are difficult to reconcile in day-to-day tasks)}
\item{Flex}{(num) degree to which employees experience freedom,
comfort, and control in how and where they work (it reflects
work-life balance, autonomy in tasks and time management)}
\item{RemoteWork}{(factor) remote-work arrangements (\code{No}: the
employee works entirely on-site; \code{Structured}: remote work on
predefined days or under fixed rules; and \code{Flexible}: remote work
without constraints)}
}
}
\usage{
data(JobEngage)
}
\description{
The data frame contains the results of a survey conducted to analyse job
engagement (the emotional commitment of employees to their organisation)
within a large company, focusing on factors previously identified through
qualitative methods (including focus groups and exploratory interviews).
\emph{Constructs} related to the possible determinants of engagement were
measured using Likert scale items, where respondents were asked to
indicate their level of agreement with a series of statements pertaining
to each factor. For example, perceived compensation fairness was measured
through items such as 'My pay reflects the responsibilities of my role'
and 'I am fairly compensated relative to others in similar jobs'. The
responses to different items referring to the same construct were then
averaged to form a composite score.
}
\references{
Piccarreta, R., Tonini, D., & Trentini, F. (2026). \emph{From Data to
Decisions. An Introduction to Applied Statistics}. BUP. ISBN
9788823824096.
}
\keyword{datasets}
