% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{MktDATA.Orig}
\alias{MktDATA.Orig}
\title{Data: MktDATA.Orig}
\format{
A data frame with 2224 observations and the following 19
variables (levels of the variables listed in alphabetical order):
#' * \strong{\code{CustId}} (num): customer's identification label
\itemize{
\item \strong{\code{Gender}} (chr): customer's gender (\code{F}, \code{M})
\item \strong{\code{Age}} (num): customer's age (in years)
\item \strong{\code{Education}} (chr): customer's level of education(\code{College}, \code{Graduate},
\item \strong{\code{Marital_Status}} (chr): customer's marital status
\item \strong{\code{Children}} (num): number of children in the household
\item \strong{\code{Kids}} (num): number of kids aged less than 12 in the
household
\item \strong{\code{Income}} (chr): customer's income (measured in classes)
\item \strong{\code{Baseline}} (num): index (from 0 to 1) assigned by the marketing dept
\item \strong{\code{LikeMost}} (chr): Most frequently bought category in the last two
\item \strong{\code{TotVal}} (num): amount spent in the last 2 years
\item \strong{\code{NPickUp_Purch}} (num): number of purchases made through company's
\item \strong{\code{NWeb_Purch}} (num): number of purchases made through company's website
\item \strong{\code{NStore_Purch}} (num): number of purchases made in a physical store
\item \strong{\code{NDeals}} (num): number of products purchases with discount
\item \strong{\code{CustClass}} (chr): customer's classification (assigned by the marketing
\item \strong{\code{PastCampaigns}} (num): number of offers accepted by the customer in the
\item \strong{\code{LastCampaign}} (num): binary variable (0/1) indicating whether (1) or
\item \strong{\code{WouldSuggest}} (chr): variable signalling whether (\code{Yes}) or not
}
}
\source{
The data set has been adapted from \href{https://www.kaggle.com/code/dmitryuarov/customers-clustering-eda}{https://www.kaggle.com/code/dmitryuarov/customers-clustering-eda}.
}
\usage{
data(MktDATA.Orig)
}
\description{
This dataset contains the variables from a survey on a set of customers
of a company operating in the retail food sector. The company sells
products from 3 major categories (referred to as A, B, C) The customers can
order and acquire products in the company physical stores, or through the
company's website (in this case, they can order on the website and pick up the
order in one store). Information is collected on customers' activity in the
last two years (observation period), as well as some information retrieved
through questionnaires or fidelity cards. During such period different
marketing strategies were adopted to improve customers' fidelization, and 5
marketing campaigns were launched; a last campaign was launched at the end of
the observation period.
}
\keyword{datasets}
