% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{MktDATA}
\alias{MktDATA}
\title{Data: MktDATA}
\format{
An object of class \code{data.frame} with 2224 rows and 26 columns.
}
\usage{
data(MktDATA)
}
\description{
A data frame with 2224 observations and 26 variables.
}
\details{
This dataset is a modification of the original link{\code{MktDATA.Orig}}
dataset and it is provided for user convenience.
}
\keyword{datasets}
