% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBStats_Main_Inference.R
\name{TEST.diffprop}
\alias{TEST.diffprop}
\title{Tests for the proportion difference}
\usage{
TEST.diffprop(
  x,
  y,
  success.x = NULL,
  success.y = NULL,
  pdiff0 = 0,
  alternative = "two.sided",
  digits = 2,
  by,
  data,
  ...
)
}
\arguments{
\item{x, y}{Quoted or unquoted strings identifying the variables whose
distribution has to be analysed. \code{x} and \code{y} can be the
name of a vector or a factor in the workspace or the name of one of
the columns in the data frame specified in the \code{data} argument.
It is possible to use a mixed specification (e.g, one vector and one
column in data).}

\item{pdiff0}{Numeric value that specifies the null hypothesis to test for
(default is 0).}

\item{alternative}{A length-one character vector specifying the direction
of the alternative hypothesis. Allowed values are \code{"two.sided"}
population mean differs from \code{pdiff0}; default), or \code{"less"}
(population mean is lower than \code{pdiff0}), or \code{"greater"}
(population mean is higher than \code{pdiff0}).}

\item{digits}{Integer value specifying the number of decimals used
to round statistics; default to 2.}

\item{by}{Quoted or unquoted string identifying the variable used to
identify the two samples whose means have to be compared. Note that
only \bold{one} between \code{y} and \code{by} can be specified. More
specifically:
\itemize{
\item \code{x}, \code{y} should be specified when data on the two sub-samples
are stored in two distinct vectors,
\item \code{x}, \code{by} should be specified when data on the two sub-samples
should be obtained by splitting \code{x} into two groups based on the two
values of the by vector; in this case, \code{x} and \code{by} must have
the same length, \code{x} must be numeric, and \code{by} can take only
two values. Note that this option is available only for \bold{independent}
samples.
}}

\item{data}{An optional data frame containing the variable to be analysed.
If not found in \code{data}, the variables are taken from the environment
from which \code{TEST.diffprop()} is called.}

\item{...}{Additional arguments to be passed to low level functions.}

\item{successes.x, }{successes.y If \code{x,y} are factors, character
vectors, or numeric non-binary vectors, success must be used to indicate
the category/value corresponding to success in the populations. These
arguments can be omitted (\code{NULL}, default) if \code{x,y} are binary
numeric vectors (taking values 0 or 1 only; in this case success is
assumed to correspond to 1) or a logical vector (in these cases success
is assumed to correspond to \code{TRUE}).}
}
\value{
A table reporting the tests for the population proportion difference.
}
\description{
\code{TEST.diffprop()} computes tests for the difference between the
proportion of successes of two populations.
}
\examples{
data(MktDATA, package = "UBStats")

# Input vectors: Numeric, based on x, y
PastCampaigns_F <- MktDATA$PastCampaigns[MktDATA$Gender == "F"]
PastCampaigns_M <- MktDATA$PastCampaigns[MktDATA$Gender == "M"]
TEST.diffprop(x = PastCampaigns_F, y = PastCampaigns_M, success.x = 0,
 pdiff0 = 0.1, alternative = "greater")

# Test for the proportion difference based on logical variables
NoChildren <- (MktDATA$Children == 0)
Deals_w_child <- MktDATA$Deals.ge50[NoChildren == FALSE]
Deals_no_child <- MktDATA$Deals.ge50[NoChildren == TRUE]
TEST.diffprop(x = Deals_w_child, y = Deals_no_child, pdiff0 = 0.2,
  alternative = "less", digits = 3)

}
\seealso{
\code{\link{CI.diffprop}()} for computing confidence intervals on
the difference between two population proportions.
}
\author{
Raffaella Piccarreta \email{raffaella.piccarreta@unibocconi.it}
}
