% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBStats_Main_Inference.R
\name{TEST.mean}
\alias{TEST.mean}
\title{Test for the mean}
\usage{
TEST.mean(
  x,
  sigma = NULL,
  mu0 = 0,
  alternative = "two.sided",
  digits = 2,
  data,
  ...
)
}
\arguments{
\item{x}{An unquoted string identifying the variable whose
distribution has to be analysed. \code{x} can be the name of a vector
or a factor in the workspace or the name of one of the columns in the
data frame specified in the \code{data} argument.}

\item{sigma}{An optional numeric value representing the population standard
deviation. If \code{NULL} it estimates the population standard using the
data.}

\item{mu0}{Numeric value that specifies the null hypothesis to test for
(default is 0).}

\item{alternative}{A length-one character vector specifying the direction
of the alternative hypothesis. Allowed values are \code{"two.sided"}
population mean differs from \code{mu0}; default), or \code{"less"}
(population mean is lower than \code{mu0}), or \code{"greater"}
(population mean is higher than \code{mu0}).}

\item{digits}{Integer value specifying the number of decimals used
to round statistics; default to 2.}

\item{data}{An optional data frame containing the variable to be analysed.
If not found in \code{data}, the variables are taken from the environment
from which \code{TEST.mean()} is called.}

\item{...}{Additional arguments to be passed to low level functions.}
}
\value{
A table reporting the test for the population mean.
}
\description{
\code{TEST.mean()} performs tests on the mean of a population.
}
\examples{
data(MktDATA, package = "UBStats")

# Bilateral test, known variance
TEST.mean(NStore_Purch, sigma = 9, mu0 = 5, alternative = "two.sided",
  data = MktDATA)

# Unilateral test,  unknown variance
TEST.mean(TotVal, mu0 = 600, alternative = "less", data = MktDATA)

}
\seealso{
\code{\link{CI.mean}()} for computing confidence intervals on the
population mean.
}
\author{
Raffaella Piccarreta \email{raffaella.piccarreta@unibocconi.it}
}
