% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBStats_Main_Inference.R
\name{TEST.prop}
\alias{TEST.prop}
\title{Test for the proportion}
\usage{
TEST.prop(
  x,
  success = NULL,
  p0 = 0.5,
  alternative = "two.sided",
  digits = 2,
  data,
  ...
)
}
\arguments{
\item{x}{An unquoted string identifying the variable whose
distribution has to be analysed. \code{x} can be the name of a vector
or a factor in the workspace or the name of one of the columns in the
data frame specified in the \code{data} argument.}

\item{success}{If \code{x} is a factor, a character vector, or a numeric
non-binary vector, success must be used to indicate the category/value
corresponding to success. The argument can be omitted (\code{NULL},
default) if \code{x} is a binary numeric vector (takes values 0 or 1 only;
in this case success is assumed to be 1) or a logical vector (in these
cases success is assumed to be \code{TRUE}).}

\item{p0}{Numeric value that specifies the null hypothesis to test for
(default is 0).}

\item{alternative}{A length-one character vector specifying the direction
of the alternative hypothesis. Allowed values are \code{"two.sided"}
population mean differs from \code{p0}; default), or \code{"less"}
(population mean is lower than \code{p0}), or \code{"greater"}
(population mean is higher than \code{p0}).}

\item{digits}{Integer value specifying the number of decimals used
to round statistics; default to 2.}

\item{data}{An optional data frame containing the variable to be analysed.
If not found in \code{data}, the variables are taken from the environment
from which \code{TEST.prop()} is called.}

\item{...}{Additional arguments to be passed to low level functions.}
}
\value{
A table reporting the tests for the population proportion of
successes.
}
\description{
\code{TEST.prop()} performs tests on the proportion of successes for a
population.
}
\examples{
data(MktDATA, package = "UBStats")

# Bilateral test (character or factor: specific value/level)
TEST.prop(WouldSuggest, success = "Yes", p0 = 0.7, data = MktDATA)

# Unilateral test (binary variable - takes values 0/1)
TEST.prop(LastCampaign, p0 = 0.2, alternative = "less", digits = 3,
  data = MktDATA)

# Unilateral test (logical variable,  taking values TRUE/FALSE)
TEST.prop(Deals.ge50, p0 = 0.13, alternative = "greater", digits = 3,
  data = MktDATA)

}
\seealso{
\code{\link{CI.prop}()} for computing confidence intervals on the
proportion of successes.
}
\author{
Raffaella Piccarreta \email{raffaella.piccarreta@unibocconi.it}
}
