% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{Time_Social}
\alias{Time_Social}
\title{Data: Time_Social}
\format{
{
A dataframe with \emph{n} = 5976 observations and the following 4 variables
(levels of the variables listed in alphabetical order):
\itemize{
\item \strong{User} (\code{int}): user id
\item \strong{Time} (\code{num}): Time (in minutes) spent on the platform during a
specific day
\item \strong{Area} (\code{chr}): geographic area where the user lives (A, B)
\item \strong{Push} (\code{chr}): variable indicating whether the user activated push
notifications (No, Yes)
}
}
}
\usage{
data(Time_Social)
}
\description{
The dataframe contains information on the time spent on a social media
platform, after the algorithm that selects displayed content to users was
modified. Specifically, attention is limited to a sample of active users
within a specific socio-demographic segment, who live in two distinct areas.
}
\keyword{datasets}
