% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{Transition}
\alias{Transition}
\title{Data: Transition}
\format{
{
A dataframe with \emph{n} = 130 observations and the following 4 variables
(levels of the variables listed in alphabetical order):
\itemize{
\item \strong{Employee} (\code{int}): employee id
\item \strong{Pre} (\code{num}): employee’s productivity with the old software
\item \strong{Post} (\code{num}): employee’s productivity with the new software
(after the training program)
\item \strong{Department} (\code{chr}): employee’s department (Dept1, Dept2)
}
}
}
\usage{
data(Transition)
}
\description{
The dataframe refers to data on the employees of a company that adopted a new
management software system to replace its legacy tools. To facilitate a
smooth transition, all administrative staff will participate in a structured
training program designed to build proficiency with the new system. As part
of the rollout, employees from two departments have been selected to complete
the training first. The dataset contains information on each
(trained) employee's productivity using the old and the new software,
recorded after a defined period of use.
}
\keyword{datasets}
