% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_package.R
\docType{package}
\name{UBStats-package}
\alias{UBStats}
\alias{UBStats-package}
\title{UBStats: A package for composite-based structural equation modeling}
\description{
Estimate, analyze, test, and study linear, nonlinear, hierarchical and
multigroup structural equation models using composite-based approaches and procedures including estimation
techniques such as partial least squares path modeling (PLS) and its derivatives
(PLSc, ordPLSc, robustPLSc), generalized structured component analysis (GSCA),
generalized structured component analysis with uniqueness terms (GSCAm),
generalized canonical correlation analysis (GCCA) unit weights (sum score) and
fixed weights, as well as several tests and typical postestimation
procedures (e.g., assess the model fit, compute direct, indirect and total effects).
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/raffaellapiccarreta/UBStats/issues}
}

}
\author{
\strong{Maintainer}: Raffaella Piccarreta \email{raffaella.piccarreta@unibocconi.it}

Other contributors:
\itemize{
  \item Sergio Venturini \email{sergio.venturini@unicatt.it} [contributor]
}

}
\keyword{internal}
