% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_package.R
\docType{package}
\name{UBStats-package}
\alias{UBStats-package}
\alias{UBStats-pkg}
\alias{UBStats-p}
\title{UBStats: A package for teaching statistics at Università Bocconi}
\description{
The \pkg{UBStats} package ...
}
\keyword{internal}
